requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
ccxt>=4.1.0
python-dotenv==1.0.0
ta==0.10.2
pandas-ta>=0.3.14b0
scikit-learn==1.3.0
tensorflow>=2.15.0
scipy==1.11.4
statsmodels>=0.14.1
python-binance==1.0.26
websocket-client==1.7.0
aiohttp>=3.9.1
asyncio>=3.4.3
streamlit==1.24.1
streamlit-autorefresh==0.0.1
yfinance==0.2.36
requests==2.31.0
TA-Lib==0.4.28
python-dotenv==1.0.0
numpy>=1.19.5
pandas>=1.3.0
yfinance>=0.1.63
tensorflow>=2.7.0
scikit-learn>=0.24.2
TA-Lib>=0.4.19
matplotlib>=3.4.3
pytrends
snscrape
beautifulsoup4
gunicorn==21.2.0
